There is a race condition between _process.HasExited and BeginGuiMacro.
BeginGuiMacro is invoked for the _process and it somehow happens that BeginGuiMacro
never comes back - perhaps the _process has exited before it could handle the macro.
At this point when the session is being disposed I see no benefit of closing it nicely, so
wait additional 10 miliseconds and kill the process if it has not already exited.